simplifyig coordinator files just a bit
